requirements.txt
# Core Flask dependencies
Flask==2.3.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-SocketIO==5.3.6
Flask-Migrate==4.0.5
Flask-WTF==1.1.1
Werkzeug==2.3.6

# WebSocket support
python-socketio==5.11.0
python-engineio==4.9.0

# Production server with async support
gunicorn==22.0.0
eventlet==0.36.1
# gevent has Python 3.13 compatibility issues, using eventlet only

# PostgreSQL database support - psycopg2-binary now works with Python 3.13
psycopg2-binary==2.9.10
SQLAlchemy==2.0.36

# Connection pooling handled by SQLAlchemy

# Performance monitoring and optimization
redis==5.0.1
celery==5.3.4

# Utilities
pytz==2023.3
python-dotenv==1.0.0
reportlab==4.0.4

# Additional performance libraries
ujson==5.8.0
orjson==3.9.9

# Memory optimization
pympler==0.9

# Alternative WSGI server (backup)
waitress==2.1.2